+++ src/components/Book/Write/BookInfo.tsx
@@ -1,7 +1,10 @@
 import React from "react";
 import { useParams } from "react-router-dom";
-import { HouseHoldStyledButton, TopWrapper } from "./index.styles";
-import useBook from "../../hooks/book/useBook";
+import {
+  HouseHoldStyledButton,
+  TopWrapper,
+} from "../../../pages/write/index.styles";
+import useBook from "../../../hooks/book/useBook";
 
 export default function BookInfo() {
   const { bookId } = useParams();
