+++ src/pages/signup/index.tsx
@@ -4,7 +4,7 @@ import Input from "../../components/Input";
 import useSignup from "../../hooks/useSignup";
 import * as S from "./index.styles";
 import PATH from "../../constants/path";
-import TopSection from "./TopSection";
+import TopSection from "../../components/Book/Signup/TopSection";
 
 export default function Signup() {
   const {
